# Only clone git repository once