--- a/cmd/apply.go
+++ b/cmd/apply.go
@@ -14,7 +14,7 @@ var (
 
 	applyCmd = &cobra.Command{
 		Use:   "apply",
-		Short: "apply checks if an updated is needed then apply the changes",
+		Short: "apply checks if an update is needed then apply the changes",
 		Run: func(cmd *cobra.Command, args []string) {
 			fmt.Printf("\n%s\n\n", strings.ToTitle("Apply"))
 
@@ -34,10 +34,10 @@ var (
 )
 
 func init() {
-	applyCmd.Flags().StringVarP(&cfgFile, "config", "c", "./updateCli.yaml", "config file (default is ./updateCli.yaml)")
-	applyCmd.Flags().StringVarP(&valuesFile, "values", "v", "", "values file use for templating (required {.tpl,.tmpl} config)")
+	applyCmd.Flags().StringVarP(&cfgFile, "config", "c", "./updateCli.yaml", "Sets config file or directory. (default: './updateCli.yaml')")
+	applyCmd.Flags().StringVarP(&valuesFile, "values", "v", "", "Sets values file uses for templating (required {.tpl,.tmpl} config)")
 
-	applyCmd.Flags().BoolVarP(&applyCommit, "commit", "", true, "Commit")
-	applyCmd.Flags().BoolVarP(&applyPush, "push", "", true, "Push changes")
-	applyCmd.Flags().BoolVarP(&applyClean, "clean", "", true, "clean working directory")
+	applyCmd.Flags().BoolVarP(&applyCommit, "commit", "", true, "Record changes to the repository, '--commit=false' (default: true)")
+	applyCmd.Flags().BoolVarP(&applyPush, "push", "", true, "Update remote refs '--push=false' (default: true)")
+	applyCmd.Flags().BoolVarP(&applyClean, "clean", "", true, "Remove updatecli working directory like '--clean=false '(default: true)")
 }