--- a/cmd/diff.go
+++ b/cmd/diff.go
@@ -12,7 +12,7 @@ var (
 
 	diffCmd = &cobra.Command{
 		Use:   "diff",
-		Short: "diff shows if targets update are needed",
+		Short: "diff shows changes",
 		Run: func(cmd *cobra.Command, args []string) {
 			fmt.Printf("\n%s\n\n", strings.ToTitle("Diff"))
 
@@ -32,7 +32,7 @@ var (
 )
 
 func init() {
-	diffCmd.Flags().StringVarP(&cfgFile, "config", "c", "./updateCli.yaml", "config file (default is ./updateCli.yaml)")
-	diffCmd.Flags().StringVarP(&valuesFile, "values", "v", "", "values file use for templating (required {.tpl,.tmpl} config)")
-	diffCmd.Flags().BoolVarP(&diffClean, "clean", "", true, "clean working directory")
+	diffCmd.Flags().StringVarP(&cfgFile, "config", "c", "./updateCli.yaml", "Sets config file or directory. (default: './updateCli.yaml')")
+	diffCmd.Flags().StringVarP(&valuesFile, "values", "v", "", "Sets values file uses for templating (required {.tpl,.tmpl} config)")
+	diffCmd.Flags().BoolVarP(&diffClean, "clean", "", true, "Remove updatecli working directory like '--clean=false '(default: true)")
 }