--- a/pkg/git/main.go
+++ b/pkg/git/main.go
@@ -6,6 +6,7 @@ import (
 	"path"
 
 	git "github.com/olblak/updateCli/pkg/git/generic"
+	"github.com/olblak/updateCli/pkg/tmp"
 )
 
 // Git contains settings to manipulate a git repository.
@@ -36,7 +37,7 @@ func (g *Git) Init(source string, name string) error {
 
 func (g *Git) setDirectory(version string) {
 
-	directory := path.Join(os.TempDir(), "updatecli", g.URL)
+	directory := path.Join(tmp.Directory, g.URL)
 
 	if _, err := os.Stat(directory); os.IsNotExist(err) {