Fix #105 

This PR improve several things:
* Add a clean parameter for the diff subcommand like `updatecli diff --clean=true`
* By default delete the temporary directory content at the end of a full run instead of per strategy run, so we only clone git repositories once
* Change --clean parameter to keep or delete the main working directory instead of per repository 

This PR drastically drastically improve performances, I tested on [jenkins-infra/charts](https://github.com/jenkins-infra/chart)

Before:
`updatecli apply --config updateCli/updateCli.d --values updateCli/values.yaml  39,12s user 6,91s system 43% cpu 1:45,92 total`
After:
`updatecli apply --config updateCli/updateCli.d  --values updateCli/values.yaml   30,29s user 3,60s system 66% cpu 51,092 total
`