@jetersen 2020-10-24
Hmm the cleaning of the directory is one speed improvement.

I however think another place is to perform `clone` only once perhaps by determinering each repository that would need a clone and than just use `git checkout $branch` to reset the working directory before each operation.