@olblak 2020-10-24
> Hmm the cleaning of the directory is one speed improvement.
> 
> I however think another place is to perform `clone` only once perhaps by determinering each repository that would need a clone and than just use `git checkout $branch` to reset the working directory before each operation.
> 
> https://github.com/olblak/updatecli/blob/4f51ca1e20eca618823f1ff207d67cf6dccd0ba5/pkg/engine/target/main.go#L88

That's a good point, thanks for the suggestion. It won't be too hard to do that and it will be useful for later stages.