@olblak 2020-10-29
Cloning the repository once requires some refactoring so I am going to merge this one and then a new one when ready